Правки 26.01
Внесла правки на основании рекомендации на гитхаб.
Сейчас все задачи отображаются на одном контроллере.
Остался открытым вопросы по задаче с UIDatePicker, UITextField и UILabel.
Я правильно понимаю,  в UIDatePicker пользователь выбирает дату, в UITextField  вводит в формате цифры часовой пояс, в UILabel отображается дата уже с учетом часового пояса.  Немного не понимаю, какой должен получиться результат. Можете привести пример, чтобы я поняла, что именно необходимо сделать?
Также по UIStepper, настроики для изменения изображения при активации кнопки я добавила, но у меня отображаются только синие квадратики.
